+++ app/src/main/AndroidManifest.xml
@@ -6,6 +6,7 @@
     <uses-permission android:name="android.permission.CAMERA" />
 
     <application
+        android:name=".FoodScanner"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -15,17 +16,16 @@
         android:theme="@style/Theme.FoodScanner"
         tools:targetApi="31">
         <activity
-            android:name=".Favorites"
+            android:name=".activities.Favorites"
             android:exported="false" />
         <activity
-            android:name=".History"
+            android:name=".activities.History"
             android:exported="false" />
         <activity
-            android:name=".Scan"
+            android:name=".activities.Scan"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
