+++ app/src/main/java/com/example/foodscanner/FoodScanner.kt
@@ -1,0 +1,11 @@
+package com.example.foodscanner
+
+import android.app.Application
+
+class FoodScanner : Application() {
+    public val scanHistory = HashSet<String>()
+
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
