+++ app/src/main/java/com/example/foodscanner/activities/Favorites.kt
@@ -1,8 +1,9 @@
-package com.example.foodscanner
+package com.example.foodscanner.activities
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.foodscanner.R
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class Favorites : AppCompatActivity() {
@@ -16,14 +17,12 @@ class Favorites : AppCompatActivity() {
             when (item.itemId) {
                 R.id.History -> {
                     val outIntent: Intent = Intent(applicationContext, History::class.java)
-                    outIntent.putExtra("ScannedCodes", intent.getStringArrayExtra("ScannedCodes"))
                     startActivity(outIntent)
                     overridePendingTransition(0, 0)
                     return@OnNavigationItemSelectedListener true
                 }
                 R.id.Scan -> {
                     val outIntent: Intent = Intent(applicationContext, Scan::class.java)
-                    outIntent.putExtra("ScannedCodes", intent.getStringArrayExtra("ScannedCodes"))
                     startActivity(outIntent)
                     overridePendingTransition(0, 0)
                     return@OnNavigationItemSelectedListener true
