+++ app/src/main/java/com/example/foodscanner/activities/Scan.kt
@@ -1,4 +1,4 @@
-package com.example.foodscanner
+package com.example.foodscanner.activities
 
 import android.Manifest
 import android.content.Intent
@@ -13,6 +13,10 @@ import androidx.camera.view.LifecycleCameraController
 import androidx.camera.view.PreviewView
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import com.example.foodscanner.FoodScanner
+import com.example.foodscanner.viewmodels.QrCodeDrawable
+import com.example.foodscanner.viewmodels.QrCodeViewModel
+import com.example.foodscanner.R
 import com.example.foodscanner.databinding.ActivityMainBinding
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.mlkit.vision.barcode.BarcodeScanner
@@ -27,14 +31,14 @@ class Scan : AppCompatActivity() {
     private lateinit var cameraExecutor: ExecutorService
     private lateinit var barcodeScanner: BarcodeScanner
 
-    private val scanHistory = HashSet<String>()
+    private lateinit var scanHistory: HashSet<String>
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         viewBinding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(viewBinding.root)
 
-        // this adds all of the scanned codes to ScanHistory
-        scanHistory.addAll(intent.getStringArrayExtra("ScannedCodes") ?: emptyArray<String>())
+        val app = application as FoodScanner
+        scanHistory = app.scanHistory;
 
         val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_navigator)
         bottomNavigationView.selectedItemId
@@ -43,14 +47,12 @@ class Scan : AppCompatActivity() {
             when (item.itemId) {
                 R.id.History -> {
                     val outIntent: Intent = Intent(applicationContext, History::class.java)
-                    outIntent.putExtra("ScannedCodes", scanHistory.toTypedArray())
                     startActivity(outIntent)
                     overridePendingTransition(0, 0)
                     return@OnNavigationItemSelectedListener true
                 }
                 R.id.fav -> {
                     val outIntent: Intent = Intent(applicationContext, Favorites::class.java)
-                    outIntent.putExtra("ScannedCodes", scanHistory.toTypedArray())
                     startActivity(outIntent)
                     overridePendingTransition(0, 0)
                     return@OnNavigationItemSelectedListener true
