+++ app/src/main/res/layout/activity_favorites.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Favorites">
+    tools:context=".activities.Favorites">
 
     <TextView
         android:layout_width="match_parent"
