+++ app/src/main/res/layout/activity_history.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".History">
+    tools:context=".activities.History">
 
     <TextView
         android:id="@+id/HistoryBar"
